requirements.txt
speechrecognition
pyttsx3
streamlit
pyaudio
google.generativeai
textwrap
os
python-dotenv
python-docx
pandas
PyPDF2